+++ shopify.app.toml
@@ -1,7 +1,7 @@
 # Learn more about configuring your app at https://shopify.dev/docs/apps/tools/cli/configuration
 
 client_id = "817c9fab0e11b2343b3f704a572d8a39"
-application_url = "https://resorts-hanging-shop-shoe.trycloudflare.com"
+application_url = "https://export-coverage-settings-silk.trycloudflare.com"
 embedded = false
 name = "gsan-003"
 handle = "gsan"
@@ -19,9 +19,9 @@ scopes = "read_customers,read_orders,unauthenticated_read_customers"
 
 [auth]
 redirect_urls = [
-  "https://resorts-hanging-shop-shoe.trycloudflare.com/auth/callback",
-  "https://resorts-hanging-shop-shoe.trycloudflare.com/auth/shopify/callback",
-  "https://resorts-hanging-shop-shoe.trycloudflare.com/api/auth/callback"
+  "https://export-coverage-settings-silk.trycloudflare.com/auth/callback",
+  "https://export-coverage-settings-silk.trycloudflare.com/auth/shopify/callback",
+  "https://export-coverage-settings-silk.trycloudflare.com/api/auth/callback"
 ]
 
 [pos]
